// files/src/react/modules/ledger/ledger-transactions/ledger-transactions.enums.ts
export enum TransactionTypes {
  TOPUP = 'TOPUP',
  TOPUP_REFUND = 'TOPUP_REFUND',
  CHARGE = 'CHARGE',
  REFUND = 'REFUND',
  PAYOUT = 'PAYOUT',
  ADJUSTMENT = 'ADJUSTMENT',
  TRANSFER = 'TRANSFER',
  FEE = 'FEE',
  SWEEP = 'SWEEP',
}

export enum FeeSettingTransactionTypes {
  TOPUP = 'TOPUP',
  TOPUP_REFUND = 'TOPUP_REFUND',
  TOPUP_BOOST = 'TOPUP_BOOST',
  TOPUP_REFUND_BOOST = 'TOPUP_REFUND_BOOST',
  CHARGE_TNG = 'CHARGE_TNG',
  REFUND_TNG = 'REFUND_TNG',
  CHARGE_BOOST = 'CHARGE_BOOST',
  REFUND_BOOST = 'REFUND_BOOST',
  PASSTHROUGH_FUEL = 'PASSTHROUGH_FUEL',
  PASSTHROUGH_FUEL_REFUND = 'PASSTHROUGH_FUEL__REFUND',
  PASSTHROUGH_STORE = 'PASSTHROUGH_STORE',
  PASSTHROUGH_STORE_REFUND = 'PASSTHROUGH_STORE_REFUND',
}

export enum TransactionSubTypes {
  TOPUP_BANK_ACCOUNT = 'TOPUP_BANK_ACCOUNT',
  TOPUP_CREDIT_CARD = 'TOPUP_CREDIT_CARD',
  TOPUP_DIGITAL_WALLET = 'TOPUP_DIGITAL_WALLET',
  CHARGE_WALLET = 'CHARGE_WALLET',
  REFUND_WALLET = 'REFUND_WALLET',
  CHARGE_CARD = 'CHARGE_CARD',
  REFUND_CARD = 'REFUND_CARD',
  CHARGE_BOOST = 'CHARGE_BOOST',
  REFUND_BOOST = 'REFUND_BOOST',
  TOPUP_REFUND_DIGITAL_WALLET = 'TOPUP_REFUND_DIGITAL_WALLET',
  TOPUP_REFUND_BANK_ACCOUNT = 'TOPUP_REFUND_BANK_ACCOUNT',
  TOPUP_REFUND_CREDIT_CARD = 'TOPUP_REFUND_CREDIT_CARD',
  ADJUSTMENT_WALLET = 'ADJUSTMENT_WALLET',
  ADJUSTMENT_LEDGER = 'ADJUSTMENT_LEDGER',
  TRANSFER_MERCHANT_BONUS_WALLET = 'TRANSFER_MERCHANT_BONUS_WALLET',
  TRANSFER_REFUND_MERCHANT_BONUS_WALLET = 'TRANSFER_REFUND_MERCHANT_BONUS_WALLET',
  TRANSFER_MERCHANT_TO_PREPAID = 'TRANSFER_MERCHANT_TO_PREPAID',
  TRANSFER_TRUST_TO_OPERATING = 'TRANSFER_TRUST_TO_OPERATING',
  SWEEP_COLLECTION = 'SWEEP_COLLECTION',
  SWEEP_OPERATING_COLLECTION = 'SWEEP_OPERATING_COLLECTION',
  FEE_IPAY88 = 'FEE_IPAY88',
  FEE_BOOST = 'FEE_BOOST',
  ADJUSTMENT_TOPUP = 'ADJUSTMENT_TOP_UP',
  ADJUSTMENT_CHARGE = 'ADJUSTMENT_CHARGE',
  ADJUSTMENT_MDR_CHARGE = 'ADJUSTMENT_MDR_CHARGE',
  ADJUSTMENT_PREFUND = 'ADJUSTMENT_PREFUND',
  ADJUSTMENT_MIGRATE_BALANCE = 'ADJUSTMENT_MIGRATE_BALANCE',
  ADJUSTMENT_OTHER = 'ADJUSTMENT_OTHER',
}

export enum LedgerTransactionStatus {
  PENDING = 'PENDING',
  SUCCEEDED = 'SUCCEEDED',
  ERRORED = 'ERRORED',
  FAILED = 'FAILED',
  CANCELLED = 'CANCELLED',
  REVERSED = 'REVERSED',
}

export enum TransactionPGVendors {
  IPAY88 = 'IPAY88',
  BOOST = 'BOOST',
  MAYBANK = 'MAYBANK',
}

export enum CardSchemes {
  VISA = 'VISA',
  MASTERCARD = 'MASTERCARD',
}

export enum PaymentTypes {
  LOCAL_CREDIT = 'LOCAL_CREDIT',
  LOCAL_DEBIT = 'LOCAL_DEBIT',
  INTERNATIONAL_CREDIT = 'INTERNATIONAL_CREDIT',
  INTERNATIONAL_DEBIT = 'INTERNATIONAL_DEBIT',
}